pyproject.toml
[project]
name = "bovie"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "nextcord>=3.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=9.0.1",
    "ruff>=0.11.8",
    "ty>=0.0.1a1",
    "types-python-dateutil>=2.9.0.20241206",
]

[tool.mypy]
plugins = ['pydantic.mypy']

[tool.ruff.lint]
extend-select = ["I"]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
